+++ CHANGELOG.md
@@ -37,6 +37,8 @@
   * `opIterator()`: Deprecated in favor of the new `deserializeOps()` generator
     function.
   * `opAssembler()`: Deprecated in favor of the new `serializeOps()` function.
+  * `mergingOpAssembler()`: Deprecated in favor of the new `squashOps()`
+    generator function (combined with `serializeOps()`).
   * `appendATextToAssembler()`: Deprecated in favor of the new `opsFromAText()`
     generator function.
   * `newOp()`: Deprecated in favor of the new `Op` class.